Support checking multiple locale files
Previously the KeyLister would only read translations from a file
with a particular name. For large applications, it's common to
split translations for the same locale across multiple files. This
enhances the KeyLister to return keys from all locale files in the
specified directory, provided the match the specified locale.