Replace ENV guard with explicit 'start' method
Previously we used to recommend an 'after(:suite)' RSpec hook to
print coverage, which caused the output to become mangled with the
test output, and get printed before the RSpec summary message, which
was confusing [1]. This introduces a 'start' method that uses an
'at_exit' hook to separate printing the report from the RSpec output.

Previously we used to hook into I18n if a specific 'I18N_COVERAGE'
environment variable was set, which is inconsistent with the need to
explicitly call the 'report' method in order to obtain the results.
This reuses the above 'start' method to add the hook into I18n, and
warns any existing users to switch to this.

[1]: example mangled output (below)
............................................I18n Coverage (etc.)

Finished in 0.85839 seconds (files took 2.5 seconds to load)
44 examples, 0 failures