Allow prefix matching of I18n keys
Previously we required a full match in order to say a key was used,
which is problematic for advanced usages of the I18n API, where a
prefix key can be used to retrieve the underlying hash for further
processing. This relaxes the matching criteria, so that we say a key
as 'used' if there was a call to 'I18n.t' with a prefix of that key.

<locales/en.yml>
en:
  home:
    some_error:
      title: Error title
      description: Error description

<app/controller.rb>
render :edit,
       error: I18n.t("home.some_error")