Enable switching the printer for the report
Previously only a textual report was printed, which can be difficult
to read if it contains a large number of unused keys. This extracts
the existing printer code into a new 'BasicPrinter' class, introduces
a new SimpleCov-style 'FilePrinter', and makes the choice of printer
configurable. Having the unused keys in a machine readable file is
useful for future analysis, outside of the codebase.