Relax coverage check for non-translated keys
Previously a key was only covered if it was used in a call to I18n#t.
This relaxes the way keys are collected to also include those that
were not stricly translated e.g. when calling I18n#exists?.

The reason for this change is to make it easier and clearer to cover
repetitive translations. Sometimes we have a view that shows a different
translation depending on a variable. It's not practical to do a UI test
with all the possible values, since UI tests are expensive to run. Instead,
we can write a test to check a translation 'exists?' for all the possible
values the variable can have. We could call I18n#t and check the result
is not nil, but that would be less explicit than 'exists?'.

Reference: https://github.com/alphagov/content-publisher/blob/f0b498527c708f021b62c36ed32833644152d633/spec/models/status_spec.rb#L5